PCRedux - quantitative PCR Data Mining and Machine Learning Helper Tool
Source: R/PCRedux.R

PCRedux package is a toolbox for the analysis of sigmoid curve (qPCR) data.

l4

Format

An object of class list of length 11.

Machine learning

In machine learning and statistics, the classification should be used to identify a new unknown observation. This observation is assigned to a number of categories. One basis is training data sets containing observations with known classes. Using the example of sigmoid amplification curves, this could be an assignment to the class "negative","ambiguous" or "positive". Basically, a number of descriptors (e. g., characteristics of curvature) are required to be able to assign classes. This package contains functions for extracting characteristics. In addition, the package contains data sets of classified amplification curves.

Examples

# Use the mblrr function to analyse amplification curves
library(qpcR)

#> Loading required package: MASS

#> Loading required package: minpack.lm

#> Loading required package: rgl

#> Loading required package: robustbase

#> Loading required package: Matrix

#> 
#> Attaching package: ‘qpcR’

#> The following objects are masked from ‘package:PCRedux’:
#> 
#>     l4, l5, l6, l7

mblrr(x=boggy[, 1], y=boggy[, 2])

#> mblrr_intercept_bg     mblrr_slope_bg       mblrr_cor_bg mblrr_intercept_pt 
#>        0.181622824        0.005868179        0.800678697        1.954160826 
#>     mblrr_slope_pt       mblrr_cor_pt 
#>       -0.003070031        0.000000000