A function to calculate numerous features from amplification curve data from a quantitative PCR experiment.

encu (ENcode CUrves) is a function to calculate numerous features of a large amplification curve data set. The pcrfit_single is performing the analysis for a single process.

encu(data, detection_chemistry = NA, device = NA)

Arguments

data
is the data set containing the cycles and fluorescence amplitudes.

detection_chemistry
contains additional meta information about the detection chemistry (e.g., probes, intercalating dye) that was used.

device
contains additional meta information about the qPCR system that was used.

Value

The output of the encu function is identical to the pcrfit_single function.

Examples

# Calculate curve features of an amplification curve data. Note that not all
# available CPU cores are used. If need set "all" to use all available cores.
# In this example the testdat data set from the qpcR package is used.
# The samples F1.1 and F1.2 are positive amplification curves. The samples
# F1.3 and F1.4 are negative.

library(qpcR)
res_encu <- encu(testdat[, 1:3])

#> Loading required package: bcp

#> Loading required package: grid

res_encu

#>   runs  cpD1  cpD2 cpD2_approx cpD2_ratio      eff   sliwin cpDdiff
#> 1 F1.1 17.49 15.62    15.29714   1.021106 1.819425 1.739444    1.87
#> 2 F1.2 17.39 15.59    15.16525   1.028008 1.782711 1.737737    1.80
#>   loglin_slope cpD2_range top      f.top tdp     f.tdp bg.stop amp.stop
#> 1    0.1382870   4.527127  10 0.01391000  33 0.9830427       9       30
#> 2    0.1421122   4.496545  10 0.01330956  29 0.9889347       9       27
#>     b_slope b_model_param c_model_param d_model_param e_model_param
#> 1 -12.73839     -13.47658   0.003242222     0.9894974      18.24647
#> 2 -13.90384     -16.36373   0.002305736     0.9830211      18.46327
#>   f_model_param f_intercept convInfo_iteratons qPCRmodel qPCRmodelRF
#> 1     0.7304393   0.9283717                  8        l7          l7
#> 2     0.5350400   0.7622996                  8        l7          l7
#>         minRFU maxRFU        init2      fluo    slope_bg intercept_bg
#> 1 -0.004237230      1 1.700487e-05 0.1953197 0.001733801 -0.006777540
#> 2 -0.003998994      1 2.631616e-05 0.2160592 0.001635822 -0.006490719
#>       sigma_bg     sd_bg head2tail_ratio mblrr_slope_pt mblrr_intercept_bg
#> 1 0.0002022220 0.4527011   -0.0007122453  -0.0003719424       -0.006442819
#> 2 0.0002437449 0.4453795   -0.0009850845  -0.0002171012       -0.006274295
#>   mblrr_slope_bg mblrr_cor_bg mblrr_intercept_pt mblrr_cor_pt   polyarea
#> 1    0.001572672    0.9323391           1.008902            0 -0.1559269
#> 2    0.001548088    0.9198655           1.001047            0 -0.1684371
#>   peaks_ratio autocorellation cp_e.agglo cp_bcp amptester_shapiro amptester_lrt
#> 1  0.02584771       0.4994275          2     14             FALSE          TRUE
#> 2  0.03090811       0.4521168          2     12             FALSE          TRUE
#>   amptester_rgt amptester_tht amptester_slt amptester_polygon
#> 1          TRUE          TRUE          TRUE          4.742807
#> 2          TRUE          TRUE          TRUE          4.782519
#>   amptester_slope.ratio hookreg_hook hookreg_hook_slope hookreg_hook_delta
#> 1             0.2223128            1       -0.001613054                 16
#> 2             0.2163902            1       -0.002352293                 19
#>   central_angle number_of_cycles detection_chemistry device
#> 1    -0.9901063               49                  NA     NA
#> 2    -0.9896863               49                  NA     NA