hookregNL - A function to calculate the slope of amplification curves in the tail region
Source: R/hookregNL.R

hookregNL is a function to calculate the slope and intercept of an amplification curve from a quantitative PCR experiment. The idea is that a strong negative slope at the end of an amplification curve is indicative for a hook effect (see Barratt and Mackay 2002). In contrast to hookreg fits this function a sex-parameter model to the amplification curve and extracts the coefficient, which describes the slope.

hookregNL(x, y, plot = FALSE, level = 0.995, simple = TRUE,
  manualtrim = 5)

Arguments

x
is the cycle numbers (x-axis).

y
is the cycle dependent fluorescence amplitude (y-axis).

plot
is a logical parameter indicating if the data should be plotted, Default: FALSE.

level
the confidence level required, Default: 0.99.

simple
is a logical parameter. If TRUE (default) only the slope, confidence interval and decisions are shown as output

manualtrim
is the number of cycles that should be reomoved from the background. (data.frame). If FALSE, a list including the 6-parameter model is the output.

References

K. Barratt, J.F. Mackay, Improving Real-Time PCR Genotyping Assays by Asymmetric Amplification, J. Clin. Microbiol. 40 (2002) 1571--1572. doi:10.1128/JCM.40.4.1571-1572.2002.

See also

pcrfit confint

Examples

# Analyze data from the boggy data set for potential hook effect like
# curvature
library(qpcR)
# has hook
res <- hookregNL(boggy[, 1], boggy[, 2])

#> Waiting for profiling to be done...

res

#>         slope      CI.low       CI.up hook
#> 1 -0.01596799 -0.01872172 -0.01348229    1

# has no hook
res <- hookregNL(boggy[, 1], boggy[, 12])

#> Waiting for profiling to be done...

res

#>          slope CI.low CI.up hook
#> 1 0.0003363785     NA    NA    0