pcrfit_single - A function to extract features from an amplification curve

The pcrfit_single is responsible for the extraction of features from amplification curve data. The function can be used for custom functions for a paralleled analysis of amplification curve data. An example is given in the vignette.

pcrfit_single(x)

Arguments

x
is the data set containing the fluorescence amplitudes.

Value

Output Description

"eff"	qPCR amplification efficiency	numeric

"cpD1"	maximum of the first derivative curve	numeric

"cpD2"	maximum of the second derivative curve	numeric

"fluo"	raw fluorescence value at the point defined by cpD2	numeric

"init2"	initial template fluorescence from an exponential model	numeric

"top"	takeoff point. When no top can be determined, the tob value is set to the first cycle number.	numeric

"f.top"	fluorescence at takeoff point. When no f.tdp can be determined, the f.tdp value is set to the RFU value at the first cycle number.	numeric

"tdp"	takes the maximum x fluorescence subtracted by reverse values of the fluorescence and calculates then the fake takeoff point. It is so to speak the take down point (tdp). When no tdp can be determined, the tdb value is set to the last cycle number.	numeric

"f.tdp"	fluorescence at tdp point. When no f.tdp can be determined, the f.tdp value is set to the RFU value at the last cycle number.	numeric

"sliwin"	PCR efficiency by the 'window-of-linearity' method	numeric

"b_slope"	Is the slope of seven parameter model	"f_intercept"

Is the intercept of the seven parameter model	"convInfo_iteratons"	Number of iterations needed to fit the model

numeric	"cpDdiff"	absolute difference between cpD1 and cpD2

numeric	"slope_bg"	slope of the first cycles

numeric	"intercept_bg"	intercept of the first cycles

numeric	"polyarea"	area of a polygon given by the vertices in the vectors cycles and fluorescence

numeric	"changepoint_e.agglo"	agglomerative hierarchical estimate for multiple change points

numeric	"changepoint_bcp"	change point by Bayesian analysis methods

numeric	"qPCRmodel"	non-linear model determined for the analysis

factor	"amptester_shapiro"	tests based on the Shapiro-Wilk normality test if the amplification curve is just noise

binary	"amptester_lrt"	performs a cycle dependent linear regression and determines if the coefficients of determination deviates from a threshold

binary	"amptester_rgt"	Resids growth test (RGt) tests if fluorescence values in a linear phase are stable

binary	"amptester_tht"	Threshold test (THt) takes the first 20 percent and the last 15 percent of any input data set and performs a Wilcoxon rank sum tests.

binary	"amptester_slt"	Signal level test compares 1. the signals by a robust "sigma" rule by median + 2 * mad and 2. by comparison of the signal/noise ratio

binary	"amptester_polygon"	pco test (pco) determines if the points in an amplification curve (like a polygon, in particular non-convex polygons) are in a "clockwise" order.

binary	"amptester_slope.ratio"	SlR uses the inder function to find the approximated first derivative maximum, second derivative minimum and the second derivative maximum. These are used for a regression analysis with the corresponding fluorescence amplitude data.

numeric	"minRFU"	minimum of fluorescence amplitude (percentile 0.01)

numeric	"maxRFU"	maximum of fluorescence amplitude (percentile 0.99)

numeric	"bg.start_norm"	takes the start (cycle) the amplification curve background based on the bg.max function and normalizes it to the total cycle number

numeric	"bg.stop"	estimates the end (cycle) the amplification curve background based on the bg.max function and normalizes it to the total cycle number

numeric	"amp.stop"	estimates the end (cycle) of the amplification curve based in the bg.max function and normalizes it to the total cycle number

numeric	"head2tail_ratio"

numeric	"autocorellation"

numeric	"mblrr_intercept_bg"

numeric	"mblrr_slope_bg"

numeric	"mblrr_cor_bg"

numeric	"mblrr_intercept_pt"

numeric	"mblrr_slope_pt"

numeric	"mblrr_cor_pt"

numeric	"amp_cor_MIC"

numeric	"hookreg_hook"	estimate of hook effect like curvature

binary	"hookreg_hook_slope"	estimate of slope of the hook effect like curvature

numeric	"peaks_ratio"	Takes the estimate approximate local minimums and maximums

"loglin_slope"	slope determined by a linear model of the data points from the minimum and maximum of the second derivative

numeric	"cpD2_range"	cycle difference between the maximum and the minimum of the second derivative curve

numeric	"sd_bg	shows the standard deviation of the fluorescence in the ground phase

numeric		"eff"

Details

Details can be found in the vignette.

References

M. Febrero-Bande, M.O. de la Fuente, others, Statistical computing in functional data analysis: The R package fda.usc, Journal of Statistical Software. 51 (2012) 1--28. http://www.jstatsoft.org/v51/i04/

A.-N. Spiess, C. Deutschmann, M. Burdukiewicz, R. Himmelreich, K. Klat, P. Schierack, S. Roediger, Impact of Smoothing on Parameter Estimation in Quantitative DNA Amplification Experiments, Clinical Chemistry. 61 (2015) 379--388. doi:10.1373/clinchem.2014.230656.

S. Roediger, A. Boehm, I. Schimke, Surface Melting Curve Analysis with R, The R Journal. 5 (2013) 37--53. http://journal.r-project.org/archive/2013-2/roediger-bohm-schimke.pdf.

S. Roediger, M. Burdukiewicz, K.A. Blagodatskikh, P. Schierack, R as an Environment for the Reproducible Analysis of DNA Amplification Experiments, The R Journal. 7 (2015) 127--150. http://journal.r-project.org/archive/2015-1/RJ-2015-1.pdf.

S. Pabinger, S. Roediger, A. Kriegner, K. Vierlinger, A. Weinhauusel, A survey of tools for the analysis of quantitative PCR (qPCR) data, Biomolecular Detection and Quantification. 1 (2014) 23--33. doi:10.1016/j.bdq.2014.08.002.

S. Roediger, M. Burdukiewicz, P. Schierack, chipPCR: an R package to pre-process raw data of amplification curves, Bioinformatics. 31 (2015) 2900--2902. doi:10.1093/bioinformatics/btv205.

See also

Examples

# Load the chipPCR package and analyze from the C126EG685 the first qPCR run
# "A01" (column 2).
library(chipPCR)
res <- pcrfit_single(C126EG685[, 2])

#> N:36, idsize:36, idval1:1, idval22:22
#> mm: 1, nn2: 36, N:36, cumksize.size: 36