A function to calculate numerous features from amplification curve data from a quantitative PCR experiment.
Source: R/encu.R

encu (ENcode CUrves) is a function to calculate numerous features of a large amplification curve data set. The pcrfit_single is performing the analysis for a single process.

encu(data, detection_chemistry = NA, device = NA)

Arguments

data
is the data set containing the cycles and fluorescence amplitudes.

detection_chemistry
contains additional meta information about the detection chemistry (e.g., probes, intercalating dye) that was used.

device
contains additional meta information about the qPCR system that was used.

Value

The output of the encu function is identical to the pcrfit_single function.

Examples

# Calculate curve features of an amplification curve data. Note that not all
# available CPU cores are used. If need set "all" to use all available cores.
# In this example the testdat data set from the qpcR package is used.
# The samples F1.1 and F1.2 are positive amplification curves. The samples
# F1.3 and F1.4 are negative.

library(qpcR)
res_encu <- encu(testdat[, 1:3])

#>    |                                                  | 0 % ~calculating     |+++++++++++++++++++++++++                         | 50% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 02s

res_encu

#>   runs  cpD1  cpD2      eff   sliwin    cpDdiff diffQ2_slope diffQ2_Cq_range
#> 1 F1.1 17.49 15.62 1.819425 1.739444 0.03816327            0      0.09127203
#> 2 F1.2 17.39 15.59 1.782711 1.737737 0.03673469            0      0.09277801
#>         top      f.top       tdp     f.tdp bg.stop_norm amp.stop_norm   b_slope
#> 1 0.2040816 0.01391000 0.5510204 0.9712161    0.1836735     0.6122449 -12.73839
#> 2 0.2040816 0.01330955 0.5714286 0.9853867    0.1836735     0.5510204 -13.90384
#>   f_intercept convInfo_iteratons qPCRmodel qPCRmodel_reverse       minRFU
#> 1   0.9283718                  8        l7                l7 -0.004237230
#> 2   0.7622996                  8        l7                l7 -0.003998994
#>   maxRFU        init2      fluo    slope_bg intercept_bg     sd_bg
#> 1      1 1.700487e-05 0.1953197 0.001581130 -0.006475498 0.4229451
#> 2      1 2.631616e-05 0.2160592 0.001655031 -0.006660901 0.4395818
#>   head_to_tail_ratio mblrr_slope_pt mblrr_intercept_bg mblrr_slope_bg
#> 1      -0.0007122453  -0.0003719424       -0.006442819    0.001572672
#> 2      -0.0009850845  -0.0002171012       -0.006274295    0.001548088
#>   mblrr_cor_bg mblrr_intercept_pt mblrr_cor_pt   polyarea peaks_min_max_ratio
#> 1    0.9323391           1.008902            0 -0.1559269          0.02573190
#> 2    0.9198655           1.001047            0 -0.1684371          0.03253091
#>   autocorellation changepoint_e.agglo changepoint_bcp amptester_shapiro
#> 1       0.5814756                   2              11             FALSE
#> 2       0.5478919                   2              13             FALSE
#>   amptester_lrt amptester_rgt amptester_tht amptester_slt amptester_polygon
#> 1          TRUE          TRUE          TRUE          TRUE          1.266341
#> 2          TRUE          TRUE          TRUE          TRUE          1.273192
#>   amptester_slope.ratio hookreg_hook hookreg_hook_slope hookreg_hook_delta
#> 1             0.2219895            1       -0.001613054                 16
#> 2             0.2167183            1       -0.002352293                 19
#>   detection_chemistry device
#> 1                  NA     NA
#> 2                  NA     NA