A function to calculate numerous features from amplification curve data from a quantitative PCR experiment.

encu (ENcode CUrves) is a function to calculate numerous features of a large amplification curve data set. The pcrfit_single is performing the analysis for a single process.

encu(data, detection_chemistry = NA, device = NA)

Arguments

data
is the data set containing the cycles and fluorescence amplitudes.

detection_chemistry
contains additional meta information about the detection chemistry (e.g., probes, intercalating dye) that was used.

device
contains additional meta information about the qPCR system that was used.

Value

The output of the encu function is identical to the pcrfit_single function.

Examples

# Calculate curve features of an amplification curve data. Note that not all
# available CPU cores are used. If need set "all" to use all available cores.
# In this example the testdat data set from the qpcR package is used.
# The samples F1.1 and F1.2 are positive amplification curves. The samples
# F1.3 and F1.4 are negative.

library(qpcR)
res_encu <- encu(testdat[, 1:3])

#> Loading required package: bcp

#> Loading required package: grid

#> N:45, idsize:45, idval1:1, idval22:22
#> mm: 1, nn2: 45, N:45, cumksize.size: 45
#> N:45, idsize:45, idval1:1, idval22:22
#> mm: 1, nn2: 45, N:45, cumksize.size: 45

res_encu

#>   runs  cpD1  cpD2      eff   sliwin cpDdiff loglin_slope cpD2_range top
#> 1 F1.1 17.49 15.62 1.819425 1.739444    1.87    0.1382870   4.368862  10
#> 2 F1.2 17.39 15.59 1.782711 1.737737    1.80    0.1421122   4.258406  10
#>        f.top tdp     f.tdp bg.stop amp.stop   b_slope f_intercept
#> 1 0.01391000  27 0.9712161       9       30 -12.73839   0.9283718
#> 2 0.01330955  28 0.9853867       9       27 -13.90384   0.7622996
#>   convInfo_iteratons qPCRmodel qPCRmodelRF       minRFU maxRFU        init2
#> 1                  8        l7          l7 -0.004237230      1 1.700487e-05
#> 2                  8        l7          l7 -0.003998994      1 2.631616e-05
#>        fluo    slope_bg intercept_bg     sd_bg head2tail_ratio mblrr_slope_pt
#> 1 0.1953197 0.001733801 -0.006777540 0.4229451   -0.0007122453  -0.0003719424
#> 2 0.2160592 0.001635822 -0.006490719 0.4395818   -0.0009850845  -0.0002171012
#>   mblrr_intercept_bg mblrr_slope_bg mblrr_cor_bg mblrr_intercept_pt
#> 1       -0.006442819    0.001572672    0.9323391           1.008902
#> 2       -0.006274295    0.001548088    0.9198655           1.001047
#>   mblrr_cor_pt   polyarea peaks_ratio autocorellation changepoint_e.agglo
#> 1            0 -0.1559269  0.02573190       0.4883720                   2
#> 2            0 -0.1684371  0.03146392       0.4402242                   2
#>   changepoint_bcp amptester_shapiro amptester_lrt amptester_rgt amptester_tht
#> 1              10             FALSE          TRUE          TRUE          TRUE
#> 2              12             FALSE          TRUE          TRUE          TRUE
#>   amptester_slt amptester_polygon amptester_slope.ratio hookreg_hook
#> 1          TRUE          1.266341             0.2219895            1
#> 2          TRUE          1.273192             0.2167183            1
#>   hookreg_hook_slope hookreg_hook_delta number_of_cycles detection_chemistry
#> 1       -0.001613054                 16               49                  NA
#> 2       -0.002352293                 19               49                  NA
#>   device
#> 1     NA
#> 2     NA